Refactor PokeStadium and utility functions for improved battle logic
- Moved damage calculation and type effectiveness logic to utility functions for better modularity and reusability.
- Updated PokeStadium component to utilize new utility functions, enhancing readability and maintainability.
- Added detailed documentation for utility functions to clarify their purpose and usage.
- Made minor styling adjustments in the PokeStadium component for consistent capitalization and improved UI.
