+++ src/utils/utils.ts
@@ -1,5 +1,13 @@
-// Helper function to get color based on pokemon type or type_id
-export function getTypeColor(type: string | number): string {
+import { PokemonMove, PokemonStats } from "../graphql/types";
+
+/**
+ * Returns a color code based on Pokemon type or type_id
+ * Maps both string type names and numeric type_ids to their corresponding colors
+ * Falls back to gray (#777777) if type is not recognized
+ * @param type - Pokemon type (string) or type_id (number)
+ * @returns Hex color code for the given type
+ */
+export const getTypeColor = (type: string | number): string => {
   const typeColors: { [key: string | number]: string } = {
     // By name
     normal: '#A8A878',
@@ -43,3 +51,79 @@ export function getTypeColor(type: string | number): string {
 
   return typeColors[type] || '#777777';
 }
+
+/**
+ * Calculates battle damage using Pokemon damage formula
+ * @param move - The move being used in the attack
+ * @param attacker - Stats of the attacking Pokemon
+ * @param defender - Stats of the defending Pokemon
+ * @param typeEffectiveness - Type effectiveness multiplier (0.5, 1, 2 etc)
+ * @returns Calculated damage (minimum 1)
+ */
+export const calculateDamage = (
+  move: PokemonMove,
+  attacker: PokemonStats,
+  defender: PokemonStats,
+  typeEffectiveness: number
+) => {
+  // Default level 50 if not specified
+  const level = attacker.level || 50;
+  // Random variance between 0.85 and 1.0
+  const variance = Math.random() * (1 - 0.85) + 0.85;
+  // Default power 50 if move has no power
+  const movePower = move.pokemon_v2_move.power || 50;
+
+  // Pokemon damage formula:
+  // ((2 * Level / 5 + 2) * Power * (Attack / Defense)) / 50 + 2
+  // Multiplied by type effectiveness and random variance
+  const damage = Math.floor(
+    ((((2 * level) / 5 + 2) * movePower * (attacker.attack / defender.defense)) / 50 + 2) *
+    typeEffectiveness *
+    variance
+  );
+
+  // Ensure minimum damage of 1
+  return Math.max(1, damage);
+};
+
+
+/**
+ * Calculates type effectiveness multiplier between move type and defender type
+ * Based on simplified Pokemon type chart focusing on key relationships:
+ * - Fire: weak vs Water (0.5x), strong vs Grass (2x)
+ * - Water: strong vs Fire (2x), weak vs Grass (0.5x)
+ * - Grass: strong vs Water (2x), weak vs Fire (0.5x)
+ * - Electric: strong vs Water (2x), no effect on Ground (0x)
+ * - Ground: strong vs Electric (2x), weak vs Flying (0.5x)
+ * - Flying: strong vs Grass (2x), weak vs Ground (0.5x)
+ *
+ * @param moveType - Type ID of the attacking move
+ * @param defenderType - Type ID of the defending Pokemon
+ * @returns Effectiveness multiplier (0, 0.5, 1, or 2)
+ */
+export const getTypeEffectiveness = (moveType: number, defenderType: number): number => {
+  const typeChart: Record<number, Record<number, number>> = {
+    10: { 12: 0.5, 11: 2 },
+    11: { 10: 2, 12: 0.5 },
+    12: { 11: 2, 10: 0.5 },
+    13: { 11: 2, 5: 0 },
+    5: { 13: 2, 3: 0.5 },
+    3: { 12: 2, 5: 0.5 },
+  };
+
+  return typeChart[moveType]?.[defenderType] || 1;
+};
+
+/**
+ * Generates a battle message based on move effectiveness and damage
+ * @param effectiveness - Type effectiveness multiplier (0-2)
+ * @param damage - Amount of damage dealt
+ * @returns Message describing the attack effectiveness
+ */
+export const getBattleMessage = (effectiveness: number, damage: number): string => {
+  if (effectiveness > 1.5) return "It's super effective!";
+  if (effectiveness < 0.8) return "It's not very effective...";
+  if (damage < 10) return "A weak hit!";
+  if (damage > 30) return "A critical hit!";
+  return "Hit!";
+};
